% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heheplot.R
\name{clown_theme}
\alias{clown_theme}
\title{Clown theme for heheplot2 object}
\usage{
clown_theme(plt)
}
\arguments{
\item{plt}{Plot.}
}
\description{
Clown theme for heheplot2 object
}
\examples{
clown_theme(heheplot2())
}
