% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heheplot.R
\name{theme_clown}
\alias{theme_clown}
\title{Clown theme for heheplot2 object}
\usage{
theme_clown(plt)
}
\arguments{
\item{plt}{Plot.}
}
\description{
Clown theme for heheplot2 object
}
\examples{
theme_clown(heheplot2())
}
